feat: enhance list item management with product and store integration
- Updated ListItem model to include ProductID and StoreID fields, allowing for better association with products and stores.
- Modified ListService and ListRepository to handle new fields in item creation and updates, ensuring proper data management.
- Enhanced RecipeHandler to support adding missing ingredients to lists, integrating with the new list item features.
- Updated API endpoints and request structures to accommodate changes in item quantity type from integer to float, improving flexibility in item management.
- Implemented migrations to update database schema for list items, products, and shopping locations, ensuring data integrity and consistency.

These changes significantly improve the functionality and user experience in managing list items, providing better integration with product and store data.

+++ backend/internal/services/list_service.go
@@ -289,7 +289,7 @@ func (s *ListService) ClearItems(ctx context.Context, user *models.User, listID 
 }
 
 // AddItemAmount adds an amount to an existing matching item or creates it.
-func (s *ListService) AddItemAmount(ctx context.Context, user *models.User, listID uuid.UUID, name string, amount int, unit, note string) (*models.ListItem, error) {
+func (s *ListService) AddItemAmount(ctx context.Context, user *models.User, listID uuid.UUID, name string, amount float64, unit, note string) (*models.ListItem, error) {
 	if err := s.requireActiveVerifiedUser(user); err != nil {
 		return nil, err
 	}
@@ -343,7 +343,7 @@ func (s *ListService) AddItemAmount(ctx context.Context, user *models.User, list
 }
 
 // RemoveItemAmount removes an amount from a matching item, deleting it at zero.
-func (s *ListService) RemoveItemAmount(ctx context.Context, user *models.User, listID uuid.UUID, name string, amount int, unit string) (*models.ListItem, bool, error) {
+func (s *ListService) RemoveItemAmount(ctx context.Context, user *models.User, listID uuid.UUID, name string, amount float64, unit string) (*models.ListItem, bool, error) {
 	if err := s.requireActiveVerifiedUser(user); err != nil {
 		return nil, false, err
 	}
